+++ kurufrec-lexicon/src/main/java/TextFrecuencier/Frecuencier/Frecuencier.java
@@ -1,4 +1,4 @@
-package TextFrecuencier.FrecuencierContext;
+package TextFrecuencier.Frecuencier;
 import java.util.List;
 import TextFrecuencier.FrecuencyStrategy.FrecuencyStrategy;
